src folder refactor
