+++ src/routes/ProductsRoutes.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const { getProduct, getProducts, createProduct, updateProduct, deleteProduct } = require('../../controllers/ProductsController.js')
+const { getProduct, getProducts, createProduct, updateProduct, deleteProduct } = require('../controllers/ProductsController.js')
 
 const productRouter = express.Router();
